requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
timm>=0.9.0
scikit-learn>=1.2.0
pandas>=2.0.0
tqdm>=4.65.0
Pillow>=10.0.0
requests>=2.31.0
